feat(station): refactor package, use lombok, remove comment
