+++ src/main/java/atdd/station/api/response/StationResponse.java
@@ -1,27 +1,23 @@
 package atdd.station.api.response;
 
-import atdd.station.repository.Station;
+import atdd.station.domain.Station;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 import java.util.StringJoiner;
 
+import static lombok.AccessLevel.PROTECTED;
+
+@NoArgsConstructor(access = PROTECTED)
+@Getter
 public class StationResponse {
 
     private Long id;
     private String name;
 
-    protected StationResponse() {}
-
     public StationResponse(Station station) {
         this.id = station.getId();
         this.name = station.getName();
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
     }
 
     @Override
